add `typeHelpers.ts` to minimize diff in all other files
