+++ package.json
@@ -2,11 +2,45 @@
   "name": "eslint-plugin-security",
   "version": "3.0.1",
   "description": "Security rules for eslint",
-  "main": "dist/index.js",
+  "main": "./dist/index.js",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist",
+    "src"
+  ],
   "type": "module",
+  "sideEffects": false,
+  "exports": {
+    ".": {
+      "bun": {
+        "types": "./dist/index.d.ts",
+        "default": "./index.ts"
+      },
+      "module-sync": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/index.js"
+      },
+      "module": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/index.js"
+      },
+      "import": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/index.js"
+      },
+      "default": {
+        "types": "./dist/index.d.cts",
+        "default": "./dist/index.cjs"
+      }
+    },
+    "./package.json": "./package.json"
+  },
   "scripts": {
-    "build": "rimraf dist/ .rolldown/ && tsdown --config-loader='unrun' --config=tsdown.config.mts",
+    "build": "npm run clean && tsdown --config-loader='unrun' --config=tsdown.config.mts",
     "changelog": "changelog eslint-plugin-security all > CHANGELOG.md",
+    "check-exports": "attw --pack .",
+    "clean": "rimraf dist/ .rolldown/",
     "cont-int": "npm test && npm run lint",
     "format": "prettier --write .",
     "lint": "npm-run-all \"lint:*\"",
@@ -14,6 +48,7 @@
     "lint:eslint-docs": "npm run update:eslint-docs -- --check",
     "lint:js": "eslint --config=\"eslint.config.ts\" .",
     "lint:js:fix": "npm run lint:js -- --fix",
+    "prepack": "npm run build",
     "release": "npx semantic-release",
     "test": "mocha test/**",
     "typecheck": "tsc -p tsconfig.json",
@@ -38,7 +73,7 @@
     "pre-commit": "lint-staged"
   },
   "lint-staged": {
-    "*.(c|m)?[tj]sx?": [
+    "*.{ts,js,tsx,jsx,mts,mjs,cts,cjs}": [
       "prettier --write",
       "eslint --fix"
     ],
@@ -49,6 +84,7 @@
     "safe-regex": "^2.1.1"
   },
   "devDependencies": {
+    "@arethetypeswrong/cli": "^0.18.2",
     "@eslint/core": "^1.0.0",
     "@eslint/js": "^9.39.1",
     "@types/estree": "^1.0.8",
@@ -67,6 +103,7 @@
     "mocha": "^11.7.5",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.7.3",
+    "rimraf": "^6.1.2",
     "rolldown": "^1.0.0-beta.52",
     "semantic-release": "^25.0.2",
     "tsdown": "^0.17.0-beta.5",
