+++ rules/detect-object-injection.ts
@@ -4,7 +4,7 @@
  */
 
 import type { Rule } from 'eslint';
-import type { Simplify } from '../utils/import-utils.ts';
+import type { Simplify } from '../utils/typeHelpers.ts';
 
 //------------------------------------------------------------------------------
 // Rule Definition
